LayoutTests/storage/indexeddb/objectstore-autoincrement.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB's IDBObjectStore auto-increment feature.");
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

function test()
{
    result = evalAndLog("webkitIndexedDB.open('Address Book')");
    verifyResult(result);
    result.onsuccess = openSuccess;
    result.onerror = unexpectedErrorCallback;
}

function openSuccess()
{
    debug("openSuccess():");
    verifySuccessEvent(event);
    window.db = evalAndLog("db = event.result");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = setVersionSuccess;
    result.onerror = unexpectedErrorCallback;
}

function setVersionSuccess()
{
    debug("setVersionSuccess():");
    verifySuccessEvent(event);
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");
    trans.onabort = unexpectedAbortCallback;
    trans.oncomplete = setVersionCompleted;

    deleteAllObjectStores(db, createObjectStore);
}

function createObjectStore()
{
    debug("createObjectStore():");
    window.store = evalAndLog("store = db.createObjectStore('StoreWithKeyPath', {keyPath: 'id', autoIncrement: true})");
    evalAndLog("db.createObjectStore('StoreWithAutoIncrement', {autoIncrement: true})");
    evalAndLog("db.createObjectStore('PlainOldStore', {autoIncrement: false})");
    var storeNames = evalAndLog("storeNames = db.objectStoreNames");

    shouldBeEqualToString("store.name", "StoreWithKeyPath");
    shouldBe("store.keyPath", "'id'");
    shouldBe("storeNames.contains('StoreWithKeyPath')", "true");
    shouldBe("storeNames.contains('StoreWithAutoIncrement')", "true");
    shouldBe("storeNames.contains('PlainOldStore')", "true");
    shouldBe("storeNames.length", "3");

    // Let the setVersion transaction complete.
}

function setVersionCompleted()
{
    debug("setVersionCompleted():");

    window.trans = evalAndLog("trans = db.transaction([], webkitIDBTransaction.READ_WRITE)");
    trans.onabort = unexpectedAbortCallback;
    trans.oncomplete = done;

    window.store = evalAndLog("store = trans.objectStore('StoreWithKeyPath')");

    debug("Insert into object store with auto increment and key path, with key in the object.");
    result = evalAndLog("store.add({name: 'Jeffersson', number: '7010', id: 3})");
    result.onsuccess = addJefferssonSuccess;
    result.onerror = unexpectedErrorCallback;
}

function addJefferssonSuccess()
{
    debug("addJefferssonSuccess():");
    verifySuccessEvent(event);
    shouldBe("event.result", "3");

    debug("Insert into object store with auto increment and key path, without key in the object.");
    result = evalAndLog("store.add({name: 'Lincoln', number: '7012'})");
    result.onsuccess = unexpectedSuccessCallback;
    result.onerror = addLincolnError;
}

function addLincolnError()
{
    debug("addLincolnError():");
    verifyErrorEvent(event);
    // FIXME: This should be implemented, but we make it an error for now.
    shouldBe("event.code", "webkitIDBDatabaseException.UNKNOWN_ERR");

    evalAndLog("event.preventDefault()");

    window.store = evalAndLog("store = trans.objectStore('StoreWithAutoIncrement')");
    debug("Insert into object store with key gen using explicit key");
    result = evalAndLog("store.add({name: 'Lincoln', number: '7012'}, 5)");
    result.onsuccess = addLincolnSuccess;
}

function addLincolnSuccess()
{
    debug("addLincolnSuccess():");
    verifySuccessEvent(event);
    shouldBe("event.result", "5");

    result = evalAndLog("store.get(5)");
    result.onsuccess = getLincolnSuccess;
    result.onerror = unexpectedErrorCallback;
}

function getLincolnSuccess()
{
    debug("getLincolnSuccess():");
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result.name", "Lincoln");
    shouldBeEqualToString("event.result.number", "7012");

    result = evalAndLog("store.put({name: 'Abraham', number: '2107'})");
    result.onsuccess = putAbrahamSuccess;
    result.onerror = unexpectedErrorCallback;
}

function putAbrahamSuccess()
{
    debug("putAbrahamSuccess():");
    verifySuccessEvent(event);
    shouldBe("event.result", "6");

    result = evalAndLog("store.get(6)");
    result.onsuccess = getAbrahamSuccess;
    result.onerror = unexpectedErrorCallback;
}

function getAbrahamSuccess()
{
    debug("getAbrahamSuccess():");
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result.name", "Abraham");
    shouldBeEqualToString("event.result.number", "2107");

    window.store = evalAndLog("store = trans.objectStore('PlainOldStore')");
    debug("Try adding with no key to object store without auto increment.");
    result = evalAndLog("store.add({name: 'Adam'})");
    result.onsuccess = unexpectedSuccessCallback;
    result.onerror = addAdamError;
}

function addAdamError()
{
    debug("addAdamError():");
    verifyErrorEvent(event);
    shouldBe("event.code", "webkitIDBDatabaseException.DATA_ERR");

    evalAndLog("event.preventDefault()");

    result = evalAndLog("store.add({name: 'Adam'}, 1)");
    result.onsuccess = addAdamSuccess;
    result.onerror = unexpectedErrorCallback;
}

function addAdamSuccess()
{
    debug("addAdamSuccess():");
    verifySuccessEvent(event);
    shouldBe("event.result", "1");
}

test();

var successfullyParsed = true;

</script>
</body>
</html>
